+++ packages/camera/camera_avfoundation/example/ios/RunnerTests/CameraPropertiesTests.m
@@ -84,9 +84,9 @@
 - (void)testFLTGetUIDeviceOrientationForString {
   XCTAssertEqual(UIDeviceOrientationPortraitUpsideDown,
                  FLTGetUIDeviceOrientationForString(@"portraitDown"));
-  XCTAssertEqual(UIDeviceOrientationLandscapeRight,
-                 FLTGetUIDeviceOrientationForString(@"landscapeLeft"));
   XCTAssertEqual(UIDeviceOrientationLandscapeLeft,
+                 FLTGetUIDeviceOrientationForString(@"landscapeLeft"));
+  XCTAssertEqual(UIDeviceOrientationLandscapeRight,
                  FLTGetUIDeviceOrientationForString(@"landscapeRight"));
   XCTAssertEqual(UIDeviceOrientationPortrait, FLTGetUIDeviceOrientationForString(@"portraitUp"));
   XCTAssertEqual(UIDeviceOrientationUnknown, FLTGetUIDeviceOrientationForString(@"unknown"));
@@ -96,9 +96,9 @@
   XCTAssertEqualObjects(@"portraitDown",
                         FLTGetStringForUIDeviceOrientation(UIDeviceOrientationPortraitUpsideDown));
   XCTAssertEqualObjects(@"landscapeLeft",
-                        FLTGetStringForUIDeviceOrientation(UIDeviceOrientationLandscapeRight));
-  XCTAssertEqualObjects(@"landscapeRight",
                         FLTGetStringForUIDeviceOrientation(UIDeviceOrientationLandscapeLeft));
+  XCTAssertEqualObjects(@"landscapeRight",
+                        FLTGetStringForUIDeviceOrientation(UIDeviceOrientationLandscapeRight));
   XCTAssertEqualObjects(@"portraitUp",
                         FLTGetStringForUIDeviceOrientation(UIDeviceOrientationPortrait));
   XCTAssertEqualObjects(@"portraitUp", FLTGetStringForUIDeviceOrientation(-1));