+++ packages/camera/camera_android_camerax/CHANGELOG.md
@@ -1,3 +1,12 @@
+## 0.5.0+21
+
+* Changes fallback resolution strategies for camera use cases to look for a higher resolution if neither the desired
+  resolution nor any lower resolutions are available.
+
+## 0.5.0+20
+
+* Implements `setZoomLevel`.
+
 ## 0.5.0+19
 
 * Implements torch flash mode.